Refactored LDBCDatagen (Main class) to be top level within Packages
